review exported function from web-api
