--- packages/web-api/examples/vanilla-esm/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "vanilla-esm",
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "node index.js"
-  }
-}
