--- packages/web-api/examples/vanilla/index.js
@@ -1,3 +1,0 @@
-const sdkNode = require('@signalwire/node')
-
-console.log('@signalwire/node', sdkNode)
