--- packages/web-api/examples/vanilla/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "vanilla",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "node index.js"
-  }
-}
